Slight cleanup
